Extract `slate-lists` transformation functions into a separate dir
