+++ packages/slate-lists/src/lib/index.ts
@@ -1,22 +1,12 @@
 export { canDeleteBackward } from './canDeleteBackward';
-export { decreaseDepth } from './decreaseDepth';
-export { decreaseListItemDepth } from './decreaseListItemDepth';
 export { getListItemsInRange } from './getListItemsInRange';
 export { getListsInRange } from './getListsInRange';
 export { getListType } from './getListType';
 export { getNestedList } from './getNestedList';
 export { getParentList } from './getParentList';
 export { getParentListItem } from './getParentListItem';
-export { increaseDepth } from './increaseDepth';
-export { increaseListItemDepth } from './increaseListItemDepth';
 export { isCursorAtStartOfListItem } from './isCursorAtStartOfListItem';
 export { isCursorInEmptyListItem } from './isCursorInEmptyListItem';
 export { isInList } from './isInList';
 export { listItemContainsText } from './listItemContainsText';
-export { mergeListWithPreviousSiblingList } from './mergeListWithPreviousSiblingList';
-export { moveListItemsToAnotherList } from './moveListItemsToAnotherList';
-export { moveListToListItem } from './moveListToListItem';
-export { setListType } from './setListType';
-export { splitListItem } from './splitListItem';
-export { unwrapList } from './unwrapList';
-export { wrapInList } from './wrapInList';
+export { pickSubtreesRoots } from './pickSubtreesRoots';
