+++ packages/slate-lists/src/transformations/index.ts
@@ -1,0 +1,11 @@
+export { decreaseDepth } from './decreaseDepth';
+export { decreaseListItemDepth } from './decreaseListItemDepth';
+export { increaseDepth } from './increaseDepth';
+export { increaseListItemDepth } from './increaseListItemDepth';
+export { mergeListWithPreviousSiblingList } from './mergeListWithPreviousSiblingList';
+export { moveListItemsToAnotherList } from './moveListItemsToAnotherList';
+export { moveListToListItem } from './moveListToListItem';
+export { setListType } from './setListType';
+export { splitListItem } from './splitListItem';
+export { unwrapList } from './unwrapList';
+export { wrapInList } from './wrapInList';
